task 14

+++ 14-binary_tree_balance.c
@@ -35,7 +35,7 @@ int binary_tree_balance(const binary_tree_t *tree)
 
 	if (!tree)
 	return (0);
-	left = binary_tree_height(tree->left);
-	right = binary_tree_height(tree->right);
+	left = (int)binary_tree_height(tree->left);
+	right = (int)binary_tree_height(tree->right);
 	return (left - right);
 }
